requirement.txt
torch
torchvision
numpy
matplotlib
pillow
opencv-python
scipy